move InvalidReferenceError to utils/exceptions
