+++ quam/utils/exceptions.py
@@ -1,0 +1,4 @@
+class InvalidReferenceError(LookupError):
+    """Exception raised when a reference cannot be resolved."""
+
+    pass
